Output full field for all ports

+++ src/pml.h
@@ -19,7 +19,7 @@ struct pml_boundary
 
 namespace pml
 {
-	pml_boundary create(float pml_box_padding);
+	pml_boundary create(double pml_box_padding);
 	std::vector<tet> get_elements(pml_boundary pml_volumes);
 }
